+++ spec/fixtures/WorkspaceSchemes/GroupLocation/Group2Location/project_in_subgroup/project_in_subgroup/ViewController.swift
@@ -1,0 +1,20 @@
+//
+//  ViewController.swift
+//  project_in_subgroup
+//
+//  Created by Albert So on 12/7/18.
+//  Copyright © 2018 Albert So. All rights reserved.
+//
+
+import UIKit
+
+class ViewController: UIViewController {
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        // Do any additional setup after loading the view, typically from a nib.
+    }
+
+
+}
+