+++ spec/workspace/group_reference_spec.rb
@@ -3,7 +3,13 @@ require File.expand_path('../../spec_helper', __FILE__)
 module Xcodeproj
   describe Workspace do
     before do
-      @group = Workspace::GroupReference.new('Group Name', 'container')
+      @group = Workspace::GroupReference.new('Group Name', 'container', 'dir1/dir2')
+    end
+
+    it 'can be initialized with just the group name and have proper defaults' do
+      group = Workspace::GroupReference.new('Group Name')
+      group.type.should == 'container'
+      group.location.should == ''
     end
 
     it 'properly implements equality comparison' do
@@ -14,7 +20,7 @@ module Xcodeproj
 
     it 'can be initialized by the XML representation' do
       node = REXML::Element.new('Group')
-      node.attributes['location'] = 'container:'
+      node.attributes['location'] = 'container:dir1/dir2'
       node.attributes['name'] = 'Group Name'
       result = Workspace::GroupReference.from_node(node)
       result.should == @group
@@ -23,7 +29,7 @@ module Xcodeproj
     it 'returns the XML representation' do
       result = @group.to_node
       result.class.should == REXML::Element
-      result.to_s.should == "<Group location='container:' name='Group Name'/>"
+      result.to_s.should == "<Group location='container:dir1/dir2' name='Group Name'/>"
     end
 
     it 'can be converted back and forth without loss of information' do